+++ i18n-embed/src/fluent.rs
@@ -10,7 +10,7 @@ use crate::{I18nAssets, I18nEmbedError, LanguageLoader};
 pub use i18n_embed_impl::fluent_language_loader;
 
 use fluent::{concurrent::FluentBundle, FluentMessage, FluentResource, FluentValue};
-use fluent_syntax::ast::Pattern;
+use fluent_syntax::ast::{self, Pattern};
 use parking_lot::RwLock;
 use std::{borrow::Cow, collections::HashMap, fmt::Debug, sync::Arc};
 use unic_langid::LanguageIdentifier;
@@ -22,14 +22,15 @@ lazy_static::lazy_static! {
     };
 }
 
-struct LocaleBundle {
-    locale: LanguageIdentifier,
+struct LanguageBundle {
+    language: LanguageIdentifier,
     bundle: FluentBundle<Arc<FluentResource>>,
+    resources: Vec<Arc<FluentResource>>,
 }
 
-impl LocaleBundle {
-    fn new(locale: LanguageIdentifier, resources: Vec<Arc<FluentResource>>) -> Self {
-        let mut bundle = FluentBundle::new(&[locale.clone()]);
+impl LanguageBundle {
+    fn new(language: LanguageIdentifier, resources: Vec<Arc<FluentResource>>) -> Self {
+        let mut bundle = FluentBundle::new(&[language.clone()]);
 
         for resource in &resources {
             if let Err(errors) = bundle.add_resource(Arc::clone(resource)) {
@@ -39,20 +40,24 @@ impl LocaleBundle {
             }
         }
 
-        Self { locale, bundle }
+        Self {
+            language,
+            bundle,
+            resources,
+        }
     }
 }
 
-impl Debug for LocaleBundle {
+impl Debug for LanguageBundle {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "LocaleBundle(locale: {})", self.locale)
+        write!(f, "LanguageBundle(language: {})", self.language)
     }
 }
 
 #[derive(Debug)]
-struct LocaleConfig {
+struct LanguageConfig {
     current_language: LanguageIdentifier,
-    locale_bundles: Vec<LocaleBundle>,
+    language_bundles: Vec<LanguageBundle>,
 }
 
 /// [LanguageLoader] implemenation for the `fluent` localization
@@ -62,7 +67,7 @@ struct LocaleConfig {
 /// ⚠️ *This API requires the following crate features to be activated: `fluent-system`.*
 #[derive(Debug)]
 pub struct FluentLanguageLoader {
-    locale_config: RwLock<LocaleConfig>,
+    language_config: RwLock<LanguageConfig>,
     domain: String,
     fallback_language: unic_langid::LanguageIdentifier,
 }
@@ -76,13 +81,13 @@ impl FluentLanguageLoader {
         domain: S,
         fallback_language: unic_langid::LanguageIdentifier,
     ) -> Self {
-        let config = LocaleConfig {
+        let config = LanguageConfig {
             current_language: fallback_language.clone(),
-            locale_bundles: Vec::new(),
+            language_bundles: Vec::new(),
         };
 
         Self {
-            locale_config: RwLock::new(config),
+            language_config: RwLock::new(config),
             domain: domain.into(),
             fallback_language,
         }
@@ -99,22 +104,22 @@ impl FluentLanguageLoader {
         message_id: &str,
         args: HashMap<&'a str, FluentValue<'a>>,
     ) -> String {
-        let config_lock = self.locale_config.read();
+        let config_lock = self.language_config.read();
 
         let args = if args.is_empty() { None } else { Some(&args) };
 
-        config_lock.locale_bundles.iter().filter_map(|locale_bundle| {
-            locale_bundle
+        config_lock.language_bundles.iter().filter_map(|language_bundle| {
+            language_bundle
                 .bundle
                 .get_message(message_id)
                 .and_then(|m: FluentMessage<'_>| m.value)
                 .map(|pattern: &Pattern<'_>| {
                     let mut errors = Vec::new();
-                    let value = locale_bundle.bundle.format_pattern(pattern, args, &mut errors);
+                    let value = language_bundle.bundle.format_pattern(pattern, args, &mut errors);
                     if !errors.is_empty() {
                         log::error!(
                             target:"i18n_embed::fluent",
-                            "Failed to format a message for locale \"{}\" and id \"{}\".\nErrors\n{:?}.",
+                            "Failed to format a message for language \"{}\" and id \"{}\".\nErrors\n{:?}.",
                             &config_lock.current_language, message_id, errors
                         )
                     }
@@ -126,7 +131,7 @@ impl FluentLanguageLoader {
             .unwrap_or_else(|| {
                 log::error!(
                     target:"i18n_embed::fluent",
-                    "Unable to find localization for locale \"{}\" and id \"{}\".",
+                    "Unable to find localization for language \"{}\" and id \"{}\".",
                     config_lock.current_language,
                     message_id
                 );
@@ -170,13 +175,15 @@ impl FluentLanguageLoader {
     /// available in any of the languages currently loaded (including
     /// the fallback language).
     pub fn has(&self, message_id: &str) -> bool {
-        let config_lock = self.locale_config.read();
+        let config_lock = self.language_config.read();
         let mut has_message = false;
 
         config_lock
-            .locale_bundles
+            .language_bundles
             .iter()
-            .for_each(|locale_bundle| has_message |= locale_bundle.bundle.has_message(message_id));
+            .for_each(|language_bundle| {
+                has_message |= language_bundle.bundle.has_message(message_id)
+            });
 
         has_message
     }
@@ -186,23 +193,52 @@ impl FluentLanguageLoader {
     /// currently loaded, including the fallback language). Returns
     /// `Some` of whatever whatever the closure returns, or `None` if
     /// no messages were found matching the `message_id`.
-    pub fn with_fluent_message<OUT, C: Fn(fluent::FluentMessage<'_>) -> OUT>(
-        &self,
-        message_id: &str,
-        closure: C,
-    ) -> Option<OUT> {
-        let config_lock = self.locale_config.read();
+    pub fn with_fluent_message<OUT, C>(&self, message_id: &str, closure: C) -> Option<OUT>
+    where
+        C: Fn(fluent::FluentMessage<'_>) -> OUT,
+    {
+        let config_lock = self.language_config.read();
 
         if let Some(message) = config_lock
-            .locale_bundles
+            .language_bundles
             .iter()
-            .filter_map(|locale_bundle| locale_bundle.bundle.get_message(message_id))
+            .filter_map(|language_bundle| language_bundle.bundle.get_message(message_id))
             .next()
         {
             Some((closure)(message))
         } else {
             None
         }
+    }
+
+    /// Runs the provided `closure` with an iterator over the messages
+    /// available for the specified `language`. There may be duplicate
+    /// messages when they are duplicated in resources applicable to
+    /// the language. Returns the result of the closure.
+    pub fn with_message_iter<OUT, C>(&self, language: &LanguageIdentifier, closure: C) -> OUT
+    where
+        C: Fn(&mut dyn Iterator<Item = &ast::Message<'_>>) -> OUT,
+    {
+        let config_lock = self.language_config.read();
+
+        let mut iter = config_lock
+            .language_bundles
+            .iter()
+            .filter(|language_bundle| &language_bundle.language == language)
+            .flat_map(|language_bundle| {
+                language_bundle.resources.iter().flat_map(|resource| {
+                    let ast: &ast::Resource<'_> = resource.ast();
+                    ast.body.iter().filter_map(|entry| match entry {
+                        ast::ResourceEntry::Entry(entry) => match entry {
+                            ast::Entry::Message(message) => Some(message),
+                            _ => None,
+                        },
+                        _ => None,
+                    })
+                })
+            });
+
+        (closure)(&mut iter)
     }
 }
 
@@ -224,7 +260,7 @@ impl LanguageLoader for FluentLanguageLoader {
 
     /// Get the language which is currently loaded for this loader.
     fn current_language(&self) -> unic_langid::LanguageIdentifier {
-        self.locale_config.read().current_language.clone()
+        self.language_config.read().current_language.clone()
     }
 
     /// Load the languages `language_ids` using the resources packaged
@@ -249,13 +285,13 @@ impl LanguageLoader for FluentLanguageLoader {
             load_language_ids.push(&self.fallback_language);
         }
 
-        let mut locale_bundles = Vec::with_capacity(language_ids.len());
+        let mut language_bundles = Vec::with_capacity(language_ids.len());
 
-        for locale in load_language_ids {
-            let (path, file) = self.language_file(&locale, i18n_assets);
+        for language in load_language_ids {
+            let (path, file) = self.language_file(&language, i18n_assets);
 
             if let Some(file) = file {
-                log::debug!(target:"i18n_embed::fluent", "Loaded language file: \"{0}\" for language: \"{1}\"", path, locale);
+                log::debug!(target:"i18n_embed::fluent", "Loaded language file: \"{0}\" for language: \"{1}\"", path, language);
 
                 let file_string = String::from_utf8(file.to_vec())
                     .map_err(|err| I18nEmbedError::ErrorParsingFileUtf8(path.clone(), err))?;
@@ -272,19 +308,19 @@ impl LanguageLoader for FluentLanguageLoader {
 
                 let mut resources = Vec::new();
                 resources.push(Arc::new(resource));
-                let locale_bundle = LocaleBundle::new(locale.clone(), resources);
+                let language_bundle = LanguageBundle::new(language.clone(), resources);
 
-                locale_bundles.push(locale_bundle);
+                language_bundles.push(language_bundle);
             } else {
-                log::debug!(target:"i18n_embed::fluent", "Unable to find language file: \"{0}\" for language: \"{1}\"", path, locale);
-                if locale == &self.fallback_language {
-                    return Err(I18nEmbedError::LanguageNotAvailable(path, locale.clone()));
+                log::debug!(target:"i18n_embed::fluent", "Unable to find language file: \"{0}\" for language: \"{1}\"", path, language);
+                if language == &self.fallback_language {
+                    return Err(I18nEmbedError::LanguageNotAvailable(path, language.clone()));
                 }
             }
         }
 
-        let mut config_lock = self.locale_config.write();
-        config_lock.locale_bundles = locale_bundles;
+        let mut config_lock = self.language_config.write();
+        config_lock.language_bundles = language_bundles;
         config_lock.current_language = current_language.clone();
         drop(config_lock);